+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Added S3ByteStorageService for storing uploaded bytes in S3-compatible buckets.
   [CATS-992] (https://opensource.ncsa.illinois.edu/jira/browse/CATS-992)
+- Added support for archiving files in Clowder and preparing an admin email if user attempts to download archived file.
+  [CATS-981] (https://opensource.ncsa.illinois.edu/jira/browse/CATS-981)
 
 ## 1.6.2 - 2019-05-23